HALLELUJAH!!! Added 'logic' to detect a win by either player.
Was going round in circles trying to figure out why the code which worked in
run.js wouldnt work in the app. The array in app consisted of strings... Was extremeley
frustrating as no error was thrown so I thought it was due to some feature of
javascript that I was unaware of. But alas, once more, it simple.

Proud of this relatively noob project.

Need to clean code up and add some styling. Maybe long term project of
adding AI? MATHS AHHHHHHHH